refactor: merge library package to action package
