--- lib/pipx-install-action/.gitignore
@@ -1,1 +1,0 @@
-dist/
