--- lib/pipx-install-action/eslint.config.js
@@ -1,17 +1,0 @@
-import eslint from "@eslint/js";
-import tseslint from "typescript-eslint";
-
-export default [
-  eslint.configs.recommended,
-  ...tseslint.configs.recommended,
-  ...tseslint.configs.stylistic,
-  {
-    ignores: [".*", "dist"],
-  },
-  {
-    files: ["**/*.test.ts"],
-    rules: {
-      "@typescript-eslint/no-explicit-any": "off",
-    },
-  },
-];
