--- lib/pipx-install-action/src/index.ts
@@ -1,1 +1,0 @@
-export { pipxInstallAction } from "./action.js";
