+++ package.json
@@ -1,28 +1,59 @@
 {
-  "name": "root",
-  "private": true,
+  "name": "pipx-install-action",
+  "version": "1.0.0",
+  "description": "Install Python packages using pipx with cache support on GitHub Actions",
+  "keywords": [
+    "action",
+    "python",
+    "install",
+    "cache",
+    "pipx"
+  ],
+  "homepage": "https://github.com/threeal/pipx-install-action/tree/main/lib/pipx-install-action#readme",
+  "bugs": {
+    "url": "https://github.com/threeal/pipx-install-action/issues",
+    "email": "alfi.maulana.f@gmail.com"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/threeal/pipx-install-action.git",
+    "directory": "lib/pipx-install-action"
+  },
+  "license": "MIT",
+  "author": "Alfi Maulana <alfi.maulana.f@gmail.com>",
   "type": "module",
-  "workspaces": [
-    "lib/*"
+  "main": "dist/lib.js",
+  "types": "dist/lib.d.ts",
+  "files": [
+    "dist/lib.js",
+    "dist/lib.d.ts"
   ],
   "scripts": {
     "build": "rollup -c",
     "format": "prettier --write --cache .",
-    "lint": "eslint"
+    "lint": "eslint",
+    "prepack": "rollup -c",
+    "test": "jest"
   },
   "dependencies": {
-    "gha-utils": "^0.2.0",
-    "pipx-install-action": "workspace:^"
+    "cache-action": "^0.2.1",
+    "catched-error-message": "^0.0.1",
+    "gha-utils": "^0.3.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.10.0",
+    "@jest/globals": "^29.7.0",
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@rollup/plugin-typescript": "^11.1.6",
+    "@types/jest": "^29.5.13",
     "@types/node": "^22.5.4",
-    "@vercel/ncc": "^0.38.1",
     "eslint": "^9.10.0",
+    "jest": "^29.7.0",
+    "jest-extended": "^4.0.2",
     "prettier": "^3.3.3",
     "rollup": "^4.21.3",
+    "rollup-plugin-ts": "^3.4.5",
+    "ts-jest": "^29.2.5",
     "tslib": "^2.7.0",
     "typescript": "^5.6.2",
     "typescript-eslint": "^8.5.0"
