+++ src/pipx/cache.test.ts
