+++ src/pipx/cache.ts
