+++ src/pipx/environment.test.ts
