+++ src/pipx/environment.ts
