+++ src/pipx/utils.test.ts
