+++ tsconfig.json
@@ -1,10 +1,12 @@
 {
   "include": ["src"],
+  "exclude": ["**/*.test.*"],
   "compilerOptions": {
     "exactOptionalPropertyTypes": true,
     "strict": true,
     "module": "node16",
     "moduleResolution": "node16",
+    "outDir": "dist",
     "esModuleInterop": true,
     "target": "es2022",
     "skipLibCheck": true
